Updated BSB with new trace ids in the events. TODO: Fixed even streaming issues.
