+++ src/plugins/events-rabbitmq/events/emit.ts
@@ -1,4 +1,4 @@
-import {Plugin} from "../plugin";
+import {Plugin} from "../index";
 import * as amqplib from "amqp-connection-manager";
 import * as amqplibCore from "amqplib";
 import {LIB, SetupChannel} from "./lib";
@@ -57,7 +57,7 @@ export class emit {
   async onEvent(
       pluginName: string,
       event: string,
-      listener: { (args: Array<any>): Promise<void> },
+      listener: { (traceId: string | undefined, args: Array<any>): Promise<void> },
   ): Promise<void> {
     const thisQueueKey = LIB.getQueueKey(
         this.plugin,
@@ -73,25 +73,24 @@ export class emit {
           await this.receiveChannel.channel.consume(
               thisQueueKey,
               async (msg: amqplibCore.ConsumeMessage) => {
-                const start = Date.now();
+                //const start = Date.now();
                 const body = msg.content.toString();
                 const bodyObj = JSON.parse(body) as Array<any>;
                 try {
-                  await SmartFunctionCallAsync(this.plugin, listener, bodyObj);
+                  await SmartFunctionCallAsync(this.plugin, listener, bodyObj.splice(0, 1)[0], bodyObj);
                   this.receiveChannel.channel.ack(msg);
-                  const time = Date.now() - start;
-                  this.log.reportStat(
-                      `eventsrec-${this.channelKey}-${pluginName}-${event}-ok`,
-                      time,
-                  );
-                }
-                catch (err: any) {
+                  // const time = Date.now() - start;
+                  // this.log.reportStat(
+                  //     `eventsrec-${this.channelKey}-${pluginName}-${event}-ok`,
+                  //     time,
+                  // );
+                } catch (err: any) {
                   this.receiveChannel.channel.nack(msg, true);
-                  const time = Date.now() - start;
-                  this.log.reportStat(
-                      `eventsrec-${this.channelKey}-${pluginName}-${event}-error`,
-                      time,
-                  );
+                  // const time = Date.now() - start;
+                  // this.log.reportStat(
+                  //     `eventsrec-${this.channelKey}-${pluginName}-${event}-error`,
+                  //     time,
+                  // );
                   this.log.error(err.toString(), {});
                 }
               },
@@ -106,6 +105,7 @@ export class emit {
   async emitEvent(
       pluginName: string,
       event: string,
+      traceId: string | undefined,
       args: Array<any>,
   ): Promise<void> {
     const thisQueueKey = LIB.getQueueKey(
@@ -129,7 +129,7 @@ export class emit {
     }
 
     if (
-        !await this.publishChannel.channel.sendToQueue(thisQueueKey, args, {
+        !await this.publishChannel.channel.sendToQueue(thisQueueKey, [traceId, ...args], {
           expiration: this.queueOpts.messageTtl,
           contentType: "string",
           appId: this.plugin.myId,
